Move nodes utils into parser
This commit moves the utils/nodes/*.go files into parser, or
parser/visualize.  This completes the deprecation of the utils subpackage.
