+++ parser/visualize/visualize_test.go
@@ -12,7 +12,7 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
-package nodes
+package visualize
 
 import (
 	"bytes"
