--- utils/doc.go
@@ -1,23 +1,0 @@
-// Copyright (c) 2020 Kevin L. Mitchell
-//
-// Licensed under the Apache License, Version 2.0 (the "License"); you
-// may not use this file except in compliance with the License.  You
-// may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-// implied.  See the License for the specific language governing
-// permissions and limitations under the License.
-
-// Package utils provides some utility functions that developers
-// utilizing ptk may find useful.  These include the FileLocation
-// implementation of Location; the UnaryOperator and BinaryOperator
-// implementations of Node, as well as AnnotatedNode; the Literal
-// function and the Prefix, Infix, and InfixR closure generators for
-// filling out an Entry; the Visualize function, which produces a
-// visualization of an abstract syntax tree; and the ChanTokenStream
-// type and the NewListTokenStream and NewAsyncTokenStream functions.
-package utils
